+++ fews_3di/tests/test_simulation.py
@@ -1,18 +1,8 @@
 from fews_3di import simulation
-from fews_3di import utils
-from pathlib import Path
+from openapi_client.exceptions import ApiException
 
+import mock
 import pytest
-
-
-TEST_DIR = Path(__file__).parent
-EXAMPLE_SETTINGS_FILE = TEST_DIR / "example_settings.xml"
-
-
-@pytest.fixture
-def example_settings():
-    settings = utils.Settings(EXAMPLE_SETTINGS_FILE)
-    return settings
 
 
 def test_smoke(example_settings):
@@ -24,3 +14,21 @@ def test_auth_fails(example_settings):
     # The example settings of course give an authentication error.
     with pytest.raises(simulation.AuthenticationError):
         threedi_simulation.login()
+
+
+def test_auth_fails_unknown(example_settings):
+    threedi_simulation = simulation.ThreediSimulation(example_settings)
+    with mock.patch("openapi_client.AuthApi.auth_token_create") as mocked:
+        mocked.side_effect = ApiException(status=500)
+        with pytest.raises(ApiException):
+            threedi_simulation.login()
+
+
+def test_auth_succeeds(example_settings):
+    threedi_simulation = simulation.ThreediSimulation(example_settings)
+    with mock.patch("openapi_client.AuthApi.auth_token_create") as mocked:
+        mock_response = mock.Mock()
+        mock_response.access = "my tokens"
+        mocked.side_effect = [mock_response]
+        threedi_simulation.login()
+        assert threedi_simulation.configuration.api_key["Authorization"] == "my tokens"