+++ fews_3di/tests/test_utils.py
@@ -1,22 +1,27 @@
-"""Tests for utils.py"""
+"""Tests for utils.py
+
+Note: the 'example_settings' pytest fixture is defined in conftest.py.
+
+"""
 from fews_3di import utils
 from pathlib import Path
 
+import datetime
 import pytest
 
 
 TEST_DIR = Path(__file__).parent
 EXAMPLE_SETTINGS_FILE = TEST_DIR / "example_settings.xml"
 WRONG_SETTINGS_FILE = TEST_DIR / "settings_without_username.xml"
+EXAMPLE_LATERAL_CSV = TEST_DIR / "example_lateral.csv"
 
 
 def test_read_settings_smoke():
     utils.Settings(EXAMPLE_SETTINGS_FILE)
 
 
-def test_read_settings_extracts_properties():
-    settings = utils.Settings(EXAMPLE_SETTINGS_FILE)
-    assert settings.username == "pietje"
+def test_read_settings_extracts_properties(example_settings):
+    assert example_settings.username == "pietje"
 
 
 def test_read_settings_missing_username():
@@ -24,19 +29,37 @@ def test_read_settings_missing_username():
         utils.Settings(WRONG_SETTINGS_FILE)
 
 
-def test_read_settings_extracts_times():
-    settings = utils.Settings(EXAMPLE_SETTINGS_FILE)
-    assert settings.start
-    assert settings.end
-    assert settings.start.day == 26
+def test_read_settings_extracts_times(example_settings):
+    assert example_settings.start
+    assert example_settings.end
+    assert example_settings.start.day == 26
 
 
-def test_read_settings_missing_date_item():
-    settings = utils.Settings(EXAMPLE_SETTINGS_FILE)
+def test_read_settings_missing_date_item(example_settings):
     with pytest.raises(utils.MissingSettingException):
-        settings._read_datetime("middle")
+        example_settings._read_datetime("middle")
 
 
-def test_read_settings_duration():
-    settings = utils.Settings(EXAMPLE_SETTINGS_FILE)
-    assert settings.duration == 352800
+def test_read_settings_duration(example_settings):
+    assert example_settings.duration == 352800
+
+
+def test_read_settings_base_dir(example_settings):
+    assert example_settings.base_dir == TEST_DIR
+
+
+def test_lateral_timeseries_smoke(example_settings):
+    utils.lateral_timeseries(EXAMPLE_LATERAL_CSV, example_settings)
+
+
+def test_lateral_timeseries_file_missing(example_settings):
+    with pytest.raises(utils.MissingFileException):
+        utils.lateral_timeseries(Path("boodschappenlijst.csv"), example_settings)
+
+
+def test_lateral_timeseries_omit_early_timestamps(example_settings):
+    # The example laterals csv has some 2020-01-21 timestamps. Check that
+    # they're omitted when we adjust the start date.
+    example_settings.start = datetime.datetime(year=2020, month=1, day=22)
+    utils.lateral_timeseries(EXAMPLE_LATERAL_CSV, example_settings)
+    # Just a smoke test atm for code coverage.