+++ CHANGES.rst
@@ -5,13 +5,15 @@ Changelog of fews-3di
 0.2 (unreleased)
 ----------------
 
-- Nothing changed yet.
+- Added lateral processing + upload.
 
 
 0.1 (2020-04-09)
 ----------------
 
 - Started copying code from the old project.
+
+- Got 3Di api connection to work, including creating an (empty) simulation.
 
 - Initial project structure created with cookiecutter and
   https://github.com/nens/cookiecutter-python-template